fix: remove element prop and non-component export
